LPS-85578 Move to more appropriate package
