+++ modules/apps/sharing/sharing-api/src/main/java/com/liferay/sharing/security/permission/SharingEntryAction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sharing.constants;
+package com.liferay.sharing.security.permission;
 
 import java.util.Objects;
